+++ Sources/Shared/Toolkit/Data/Resource/TransformingResource.swift
@@ -29,10 +29,6 @@ open class TransformingResource: Resource {
         await _transform!(data)
     }
 
-    open func close() {
-        resource.close()
-    }
-
     // As the resource is transformed, we can't use the original source URL
     // as reference.
     public let sourceURL: AbsoluteURL? = nil