+++ Sources/Shared/Toolkit/File/FileResource.swift
@@ -14,18 +14,6 @@ public actor FileResource: Resource, Loggable {
         fileURL = file
     }
 
-    public nonisolated func close() {
-        Task { await doClose() }
-    }
-
-    private func doClose() async {
-        do {
-            try _handle?.getOrNil()?.close()
-        } catch {
-            log(.error, error)
-        }
-    }
-
     public nonisolated var sourceURL: AbsoluteURL? { fileURL }
 
     private var _length: ReadResult<UInt64?>?
@@ -55,7 +43,8 @@ public actor FileResource: Resource, Loggable {
     public func stream(range: Range<UInt64>?, consume: @escaping (Data) -> Void) async -> ReadResult<Void> {
         await handle().flatMap { handle in
             do {
-                if let range = range {
+                if var range = range {
+                    range = range.clampedToInt()
                     try handle.seek(toOffset: UInt64(max(0, range.lowerBound)))
                     if let data = try handle.read(upToCount: Int(range.upperBound - range.lowerBound)) {
                         consume(data)